Update GROMACS
Updated to latest release.

Required hwloc, as this will be wanted whenever arm support is worked out.

Customized version string so that reading the GROMACS log file will show that macports organized the build. That is potentially valuable information for GROMACS users and developers troubleshooting their runs.

Removed comment about boost, which is unused.

Removed libxml2 requirement, which is unused.